pyproject.toml
[project]
name = "dynamic-dns-updater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudflare>=4.3.1",
    "python-dotenv>=1.2.0",
    "requests>=2.32.5",
    "ruamel-yaml>=0.18.16",
]
